+++ CHANGELOG.md
@@ -19,7 +19,7 @@
 
 #### Classes
 #### Functions & Properties
-
+* !!! ProcessQueueCommandTest->queueRepository
 
 ## Crawler 11.0.4
 Crawler 11.0.4 was released on February 11th, 2022